MSA/requirements-standalone.txt
torch
transformers
soundfile
torchaudio
safetensors
pandas
Pillow
# Optional later:
# fairseq
# opencv-python
